Fix invoice calculation: Use parseFloat for decimal quantities
- Fixed materialQnty, labourQnty, foamQnty, paintingQnty calculations
- Changed parseInt to parseFloat to preserve decimal values (e.g., 4.5 instead of 4)
- Updated all calculation functions in orderCalculations.js
- Fixed invoice page calculation logic
- Cleaned up unused backup/broken files
- Items Subtotal now shows correct amounts with decimal quantities

+++ src/admin/pages/StatusManagement/StatusManagementPage.js
@@ -709,10 +709,10 @@ const StatusManagementPage = () => {
     let total = 0;
     if (order.furnitureData?.groups) {
       order.furnitureData.groups.forEach(group => {
-        total += (parseFloat(group.materialPrice) || 0) * (parseInt(group.materialQnty) || 0);
-        total += (parseFloat(group.labourPrice) || 0) * (parseInt(group.labourQnty) || 0);
+        total += (parseFloat(group.materialPrice) || 0) * (parseFloat(group.materialQnty) || 0);
+        total += (parseFloat(group.labourPrice) || 0) * (parseFloat(group.labourQnty) || 0);
         if (group.foamEnabled || group.foamPrice > 0) {
-          total += (parseFloat(group.foamPrice) || 0) * (parseInt(group.foamQnty) || 0);
+          total += (parseFloat(group.foamPrice) || 0) * (parseFloat(group.foamQnty) || 0);
         }
       });
     }
